Add top level tests
Move the simd-only tests to the top level

This allows testing even when sse4.2 isn't enabled: when it is
available, it will still test the simd implementation, but will test the
fallback otherwise.
